Updates to swift 4.1
