+++ Package.swift
@@ -1,6 +1,3 @@
-// swift-tools-version:4.0
-// The swift-tools-version declares the minimum version of Swift required to build this package.
-
 //
 //  Package.swift
 //  SSLService
@@ -19,47 +16,27 @@
 //     See the License for the specific language governing permissions and
 //     limitations under the License.
 //
-
 import PackageDescription
 
-#if os(Linux) || os(macOS) || os(iOS) || os(tvOS) || os(watchOS)
-
-var packageDependencies: [Package.Dependency] = [.package(url: "https://github.com/IBM-Swift/BlueSocket.git", from: "0.12.0")]
-var targetDependencies: [Target.Dependency] = [.byNameItem(name: "Socket")]
-
-#if os(Linux)
-packageDependencies.append(.package(url: "https://github.com/IBM-Swift/OpenSSL.git", from: "0.3.0"))
-targetDependencies.append(.byNameItem(name: "OpenSSL"))
-#endif
+#if os(Linux) || os(macOS) || os(iOS) || os(tvOS)
 
 let package = Package(
     name: "SSLService",
-
-    products: [
-        // Products define the executables and libraries produced by a package, and make them visible to other packages.
-        .library(
-            name: "SSLService",
-            targets: ["SSLService"]),
+    targets: [Target(name: "SSLService")],
+    dependencies: [
+        .Package(url: "https://github.com/IBM-Swift/BlueSocket.git", majorVersion: 0, minor: 12),
         ],
+    exclude: ["SSLService.xcodeproj", "README.md", "Sources/Info.plist"])
 
-    dependencies:
-        // Dependencies declare other packages that this package depends on.
-        // .package(url: /* package url */, from: "1.0.0"),
-        packageDependencies,
+#if os(Linux)
 
-    targets: [
-        // Targets are the basic building blocks of a package. A target defines a module or a test suite.
-        // Targets can depend on other targets in this package, and on products in packages which this package depends on.
-        .target(
-            name: "SSLService",
-            dependencies: targetDependencies,
-            exclude: ["SSLService.xcodeproj", "README.md", "Sources/Info.plist"]),
-        ]
-)
+package.dependencies.append(
+    .Package(url: "https://github.com/IBM-Swift/OpenSSL.git", majorVersion: 0, minor: 3))
+
+#endif
 
 #else
 
 fatalError("Unsupported OS")
 
 #endif
-
